#321 Separate update_ensembl_assembly management command.
