+++ rnacentral/portal/management/commands/update_ensembl_genome_mapping.py
@@ -92,9 +92,7 @@ def get_ensembl_insdc_mapping(cursor, database):
 
 
 def store_ensembl_insdc_mapping(mapping, ensembl_metadata):
-    """
-    Delete existing objects for the same assembly and insert new ones.
-    """
+    """Delete existing objects for the same assembly and insert new ones."""
     assembly_id = ensembl_metadata['assembly.default']
     EnsemblInsdcMapping.objects.filter(assembly_id=assembly_id).delete()
     data = []
@@ -110,81 +108,18 @@ def store_ensembl_insdc_mapping(mapping, ensembl_metadata):
     EnsemblInsdcMapping.objects.bulk_create(data)
 
 
-def get_ensembl_metadata(cursor, database):
-    """
-    Get metadata about genomic assemblies used in Ensembl.
-    """
-    cursor.execute("USE %s" % database)
-    sql = """
-    SELECT meta_key, meta_value
-    FROM meta
-    WHERE meta_key IN (
-    	'assembly.accession',
-    	'assembly.default',
-    	'assembly.long_name',
-    	'assembly.name',
-    	'assembly.ucsc_alias',
-    	'species.division',
-    	'species.production_name',
-    	'species.taxonomy_id',
-    	'species.common_name',
-    	'species.scientific_name',
-        'species.url',
-        'species.division'
-    )
-    """
-    cursor.execute(sql)
-    metadata = {}
-    for result in cursor.fetchall():
-        metadata[result['meta_key']] = result['meta_value']
-    return metadata
-
-
-def store_ensembl_metadata(metadata):
-    """
-    Delete existing Ensembl assembly for the same NCBI taxid and create a new one.
-    """
-    EnsemblAssembly.objects.filter(taxid=metadata['species.taxonomy_id']).delete()
-    assembly = EnsemblAssembly(
-        assembly_id=metadata['assembly.default'],
-        assembly_full_name=metadata['assembly.name'],
-        gca_accession=metadata['assembly.accession'] if 'assembly.accession' in metadata else None,
-        assembly_ucsc=metadata['assembly.ucsc_alias'] if 'assembly.ucsc_alias' in metadata else None,
-        common_name=metadata['species.common_name'],
-        taxid=metadata['species.taxonomy_id'],
-        ensembl_url=metadata['species.url'],
-        division=metadata['species.division'],
-    )
-    assembly.save()
-
-    line = "{assembly_id}\t{assembly_full_name}\t{GCA_accession}\t{assembly_ucsc}\t{common_name}\t{taxid}".format(
-        assembly_id=metadata['assembly.default'],
-        GCA_accession=metadata['assembly.accession'] if 'assembly.accession' in metadata else '',
-        assembly_full_name=metadata['assembly.name'],
-        assembly_ucsc=metadata['assembly.ucsc_alias'] if 'assembly.ucsc_alias' in metadata else '',
-        division=metadata['species.division'],
-        taxid=metadata['species.taxonomy_id'],
-        common_name=metadata['species.common_name'],
-    )
-    print line
-
-
 class Command(BaseCommand):
     """
     Usage:
     python manage.py update_ensembl_genome_mapping
     """
     def handle(self, *args, **options):
-        """
-        Main function, called by django.
-        """
+        """Main function, called by django."""
         connection = get_ensembl_connection()
         try:
             with connection.cursor() as cursor:
                 databases = get_ensembl_databases(cursor)
                 for database in databases:
-                    ensembl_metadata = get_ensembl_metadata(cursor, database)
-                    store_ensembl_metadata(ensembl_metadata)
                     mapping = get_ensembl_insdc_mapping(cursor, database)
                     store_ensembl_insdc_mapping(mapping, ensembl_metadata)
         finally:
